Move SonarLanguageRecognizer to IssueViz
