+++ src/Integration.Vsix.UnitTests/CFamily/CFamilyReproducerCommandTests.cs
@@ -29,6 +29,7 @@ using Moq;
 using SonarLint.VisualStudio.Core.Analysis;
 using SonarLint.VisualStudio.Core.CFamily;
 using SonarLint.VisualStudio.Integration.UnitTests;
+using SonarLint.VisualStudio.IssueVisualization.Editor.LanguageDetection;
 using ThreadHelper = SonarLint.VisualStudio.Integration.UnitTests.ThreadHelper;
 
 namespace SonarLint.VisualStudio.Integration.Vsix.CFamily.UnitTests
