+++ src/Integration.Vsix.UnitTests/CFamily/PreCompiledHeaders/PreCompiledHeadersEventListenerTests.cs
@@ -34,6 +34,7 @@ using SonarLint.VisualStudio.Core.CFamily;
 using SonarLint.VisualStudio.Integration.Vsix;
 using SonarLint.VisualStudio.Integration.Vsix.CFamily;
 using SonarLint.VisualStudio.Integration.Vsix.Helpers.DocumentEvents;
+using SonarLint.VisualStudio.IssueVisualization.Editor.LanguageDetection;
 
 namespace SonarLint.VisualStudio.Integration.UnitTests.CFamily
 {
