+++ src/Integration.Vsix/CFamily/PreCompiledHeaders/PreCompiledHeadersEventListener.cs
@@ -25,6 +25,7 @@ using System.Linq;
 using SonarLint.VisualStudio.Core;
 using SonarLint.VisualStudio.Core.CFamily;
 using SonarLint.VisualStudio.Integration.Vsix.Helpers.DocumentEvents;
+using SonarLint.VisualStudio.IssueVisualization.Editor.LanguageDetection;
 
 namespace SonarLint.VisualStudio.Integration.Vsix.CFamily
 {
