+++ src/IssueViz.UnitTests/Editor/LanguageDetection/SonarLanguageRecognizerTests.cs
@@ -24,13 +24,13 @@ using System.ComponentModel.Composition.Primitives;
 using System.Linq;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Utilities;
 using Moq;
 using SonarLint.VisualStudio.Core.Analysis;
-using SonarLint.VisualStudio.Integration.Vsix;
+using SonarLint.VisualStudio.Integration.UnitTests;
+using SonarLint.VisualStudio.IssueVisualization.Editor.LanguageDetection;
 
-namespace SonarLint.VisualStudio.Integration.UnitTests
+namespace SonarLint.VisualStudio.IssueVisualization.UnitTests.Editor.LanguageDetection
 {
     [TestClass]
     public class SonarLanguageRecognizerTests
